requirements.txt
Django==2.2.2
psycopg2==2.7.4
djangorestframework==3.9.4
markdown==3.1.1
django-filter==2.1.0
django-cors-headers==3.0.2
python-slugify==3.0.2
django-rest-swagger==2.2.0
GitPython==2.1.13
celery==4.3.0
